+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.0.0-rc.37",
+  sdkVersion: "4.0.0-rc.38",
   genVersion: "2.755.6",
-  userAgent: "speakeasy-sdk/typescript 4.0.0-rc.37 2.755.6 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.0.0-rc.38 2.755.6 2.0 @orq-ai/node",
 } as const;