+++ packages/orq-rc/RELEASES.md
@@ -5039,3 +5039,13 @@ Based on:
 - [typescript v4.0.0-rc.37] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.37 - packages/orq-rc
+
+## 2025-11-17 07:35:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.658.1 (2.755.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.38] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.38 - packages/orq-rc